+++ Help_en.php
@@ -11,13 +11,10 @@
  * The Catalog should only reference the Help_en.php file
  * and detect the `_help` function / source keyword.
  *
- * @uses Heredoc syntax
  * @package RosarioSIS
  * @subpackage Help
  *
  * @author François Jacquet
- *
- * @see  http://php.net/manual/en/language.types.string.php#language.types.string.syntax.heredoc
  */
 
 /**
@@ -27,32 +24,50 @@
  * @since  3.9
  *
  * @param  string $text      Text to translate.
- * @param  string $domain    Gettext domain, defaults to 'help'.
+ * @param  string $domain    Gettext domain, defaults to 'help'. For add-ons, use the module / plugin name / folder.
  * @return string Translated help text.
  */
 function _help( $text, $domain = 'help' )
 {
 	global $LocalePath;
 
-	/**
-	 * @var array
-	 */
 	static $domains_bound = array();
 
-	if ( mb_strpos( $domain, 'help' ) )
+	$addon = $domain;
+
+	if ( $domain !== 'help'
+		&& ! mb_strpos( $domain, 'help' ) )
 	{
 		$domain .= '_help';
 	}
 
 	if ( empty( $domains_bound[$domain] ) )
 	{
+		$locale_path = $LocalePath;
+
+		if ( $addon !== 'help' )
+		{
+			$locale_path = 'modules/' . $addon . '/locale';
+
+			if ( ! file_exists( $locale_path ) )
+			{
+				// Is plugin?
+				$locale_path = 'plugins/' . $addon . '/locale';
+
+				if ( ! file_exists( $locale_path ) )
+				{
+					return $text;
+				}
+			}
+		}
+
 		// Binds the messages domain to the locale folder.
-		bindtextdomain( $domain, $LocalePath );
+		bindtextdomain( $domain, $locale_path );
 
 		if ( function_exists( '_bindtextdomain' ) )
 		{
 			// Correctly bind domain when MoTranslator is in use.
-			_bindtextdomain( $domain, $LocalePath );
+			_bindtextdomain( $domain, $locale_path );
 		}
 
 		// Ensures text returned is utf-8, quite often this is iso-8859-1 by default.
@@ -61,7 +76,7 @@ function _help( $text, $domain = 'help' )
 		$domains_bound[$domain] = true;
 	}
 
-	return dgettext( 'help', $text );
+	return dgettext( $domain, $text );
 }
 
 // DEFAULT.